+++ CHANGELOG.md
@@ -5,12 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## "master" branch
+## [0.5.3] - 2024-08-16
+
+### Fixed
+- Typo: `node_coorinates_unit_cube` -> `node_coordinates_unit_cube`.
+- Bug in `nullspace_linear_elasticity`.
+- Bug in `PVector` when working in split format.
+
+### Deprecated
+
+- `near_nullspace_linear_elasticity` in favor of `nullspace_linear_elasticity`.
+
+## [0.5.2] - 2024-08-13
 
 ### Added
 
 - Split format support for `PVector`.
 - Helper functions to build partitioned sparse matrices and vectors in split format, `pvector_from_split_blocks` and `psparse_from_split_blocks`.
+- Gallery function `linear_elasticity_fem`.
+- Function `global_to_owner`.
 
 ### Deprecated